requirements.txt
python-telegram-bot==20.7
flask
yt-dlp
httpx
